change some data types of tables for better data allocation

+++ database/migrations/2023_04_27_145330_create_restaurants_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained('users', 'id')->onUpdate('cascade')->onDelete('cascade');
             $table->foreignId('type_id')->constrained('restaurant_categories', 'id')->onUpdate('cascade')->onDelete('cascade');
             $table->string('address');
-            $table->bigInteger('phone')->unsigned();
+            $table->integer('phone')->unsigned();
             $table->bigInteger('account')->unsigned();
             $table->string('name');
             $table->timestamps();
